[task_2] created package
